Move logging to a separate function
